Fix process events for testing locally
When running elastalert-test-rule with local data, the events that are processed are not updated correctly by the `process_hits` function. This function puts the event's content in a '_source' attribute, but lacks to put in place a timestamp attribute for each entry. Extracting that timestamp from said source attribute fixes the issue.