Add environment variable substitution to use_kibana4_dashboard
Without this, the Kibana host is hardcoded in the rule definition and some kind
of external rule templating is required to share the rule across multiple
deployements.  This change allows doing the same without a templating engine:
all we need is for the caller to point environment variables of their choice to
the right location when running ElastAlert.