hipchat: limit HTML message size to 10000 characters
This is a HipChat limitation, see https://www.hipchat.com/docs/apiv2/method/send_room_notification

Before, it was replacing all `\n` by `<br />` *after* checking and
restricting the size to fit within 10000 characters ... which will
definitely makes the message goes over that limit if it has been
truncated before already.

The truncation is not very nice as it may cut right in the middle of a
newly replaced `<br />`.
Something more complex could be implemented to handle that case more
nicely, but it should be good enough for now.